+++ src/extra/snappy.cpp
@@ -21,7 +21,7 @@ struct Compress {
     size_t outputLen;
     snappy::RawCompress((char *)input.payload.bytesValue,
                         input.payload.bytesSize, &_buffer[0], &outputLen);
-    return Var((uint8_t *)&_buffer[0], int64_t(outputLen));
+    return Var((uint8_t *)&_buffer[0], uint32_t(outputLen));
   }
 };
 
@@ -43,7 +43,7 @@ struct Decompress {
                           input.payload.bytesSize, &_buffer[0]);
     // easy fix for null term strings
     _buffer[len] = 0;
-    return Var((uint8_t *)&_buffer[0], int64_t(len));
+    return Var((uint8_t *)&_buffer[0], uint32_t(len));
   }
 };
